Put AddressExtender in its own .cpp and .h files
